Fixed regression targets and classification embeddings
Regression targets were using x_filled so were all 0.0. Needed to also pass x_original (with NaNs as 0.0) to the model to get actual targets. Additionally the value embeddings were only being used for the non-NaN values so the model was just instantly learning their positions and getting no loss on classification head and so loss effect on regression was very slow moving. So changed value embedding to all positions (theory is now it learns 0.0 as the sort of 'mask' token.